refactor: Rename totalInvestment to totalMoneyDeployed
Updates the interface and service logic to rename `totalInvestment` to `totalMoneyDeployed`. This change aims to provide a more accurate and descriptive name for the metric tracking the capital allocated by a team. The Google Sheets parsing logic has also been updated to be more robust in identifying this column by searching for a broader set of keywords.